fix startup without attendance credentials
