+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "bunkialo",
   "main": "expo-router/entry",
-  "version": "1.3.15",
+  "version": "1.3.16",
   "scripts": {
     "start": "expo start",
     "reset-project": "node ./scripts/reset-project.js",
